requirements.txt
PyYAML==3.11
SQLAlchemy==0.9.8
lxml==2.2.4
mysql-python==1.2.5
pyScss==1.3.4
python-daemon==1.5.2
python-ldap==2.4.13
tornado==2.4.1
xmpppy==0.5.0rc1
